WIP working through steel
