+++ examples/01_onshore_steel_mn/tech_config.yaml
@@ -67,102 +67,100 @@ technologies:
       shared_parameters:
         capacity_factor: 0.9
         plant_capacity_mtpy: 1000000.
-
       cost_parameters:
         operational_year: 2035
         o2_heat_integration: false
         lcoh: 7.37
-        inflation_rate:
         installation_time: 36 # months
         inflation_rate: 0.0 # 0 for nominal analysis
-        feedstocks:
-          oxygen_market_price: 0.0 # 0.03 $/kgO2 if `o2_heat_integration` == 1
-          excess_oxygen: 395
-          lime_unitcost: 122.1
-          carbon_unitcost: 236.97
-          electricity_cost: 48.92
-          iron_ore_pellet_unitcost: 207.35
-          oxygen_market_price: 0.03
-          raw_water_unitcost: 0.59289
-          iron_ore_consumption: 1.62927
-          raw_water_consumption: 0.80367
-          lime_consumption: 0.01812
-          carbon_consumption: 0.0538
-          hydrogen_consumption: 0.06596
-          natural_gas_consumption: 0.71657
-          electricity_consumption: 0.5502
-          slag_disposal_unitcost: 37.63
-          slag_production: 0.17433
-          maintenance_materials_unitcost: 7.72
-          natural_gas_prices:
-            "2035": 3.76232
-            "2036": 3.776032
-            "2037": 3.812906
-            "2038": 3.9107960000000004
-            "2039": 3.865776
-            "2040": 3.9617400000000003
-            "2041": 4.027136
-            "2042": 4.017166
-            "2043": 3.9715339999999997
-            "2044": 3.924314
-            "2045": 3.903287
-            "2046": 3.878192
-            "2047": 3.845413
-            "2048": 3.813366
-            "2049": 3.77735
-            "2050": 3.766164
-            "2051": 3.766164
-            "2052": 3.766164
-            "2053": 3.766164
-            "2054": 3.766164
-            "2055": 3.766164
-            "2056": 3.766164
-            "2057": 3.766164
-            "2058": 3.766164
-            "2059": 3.766164
-            "2060": 3.766164
-            "2061": 3.766164
-            "2062": 3.766164
-            "2063": 3.766164
-            "2064": 3.766164
-        finances:
-          # plant_life: 30
-          grid_prices:
-            "2035": 89.42320514456621
-            "2036": 89.97947569251141
-            "2037": 90.53574624045662
-            "2038": 91.09201678840184
-            "2039": 91.64828733634704
-            "2040": 92.20455788429224
-            "2041": 89.87291235917809
-            "2042": 87.54126683406393
-            "2043": 85.20962130894978
-            "2044": 82.87797578383562
-            "2045": 80.54633025872147
-            "2046": 81.38632144593608
-            "2047": 82.22631263315068
-            "2048": 83.0663038203653
-            "2049": 83.90629500757991
-            "2050": 84.74628619479452
-            "2051": 84.74628619479452
-            "2052": 84.74628619479452
-            "2053": 84.74628619479452
-            "2054": 84.74628619479452
-            "2055": 84.74628619479452
-            "2056": 84.74628619479452
-            "2057": 84.74628619479452
-            "2058": 84.74628619479452
-            "2059": 84.74628619479452
-            "2060": 84.74628619479452
-            "2061": 84.74628619479452
-            "2062": 84.74628619479452
-            "2063": 84.74628619479452
-            "2064": 84.74628619479452
-
-          # Additional parameters passed to ProFAST
-          financial_assumptions:
-            "total income tax rate": 0.2574
-            "capital gains tax rate": 0.15
-            "leverage after tax nominal discount rate": 0.10893
-            "debt equity ratio of initial financing": 0.624788
-            "debt interest rate": 0.050049
+        # Feedstock parameters (flattened)
+        excess_oxygen: 395
+        lime_unitcost: 122.1
+        lime_transport_cost: 0.0
+        carbon_unitcost: 236.97
+        carbon_transport_cost: 0.0
+        electricity_cost: 48.92
+        iron_ore_pellet_unitcost: 207.35
+        iron_ore_pellet_transport_cost: 0.0
+        oxygen_market_price: 0.03
+        raw_water_unitcost: 0.59289
+        iron_ore_consumption: 1.62927
+        raw_water_consumption: 0.80367
+        lime_consumption: 0.01812
+        carbon_consumption: 0.0538
+        hydrogen_consumption: 0.06596
+        natural_gas_consumption: 0.71657
+        electricity_consumption: 0.5502
+        slag_disposal_unitcost: 37.63
+        slag_production: 0.17433
+        maintenance_materials_unitcost: 7.72
+        natural_gas_prices:
+          "2035": 3.76232
+          "2036": 3.776032
+          "2037": 3.812906
+          "2038": 3.9107960000000004
+          "2039": 3.865776
+          "2040": 3.9617400000000003
+          "2041": 4.027136
+          "2042": 4.017166
+          "2043": 3.9715339999999997
+          "2044": 3.924314
+          "2045": 3.903287
+          "2046": 3.878192
+          "2047": 3.845413
+          "2048": 3.813366
+          "2049": 3.77735
+          "2050": 3.766164
+          "2051": 3.766164
+          "2052": 3.766164
+          "2053": 3.766164
+          "2054": 3.766164
+          "2055": 3.766164
+          "2056": 3.766164
+          "2057": 3.766164
+          "2058": 3.766164
+          "2059": 3.766164
+          "2060": 3.766164
+          "2061": 3.766164
+          "2062": 3.766164
+          "2063": 3.766164
+          "2064": 3.766164
+        # Financial parameters (flattened)
+        grid_prices:
+          "2035": 89.42320514456621
+          "2036": 89.97947569251141
+          "2037": 90.53574624045662
+          "2038": 91.09201678840184
+          "2039": 91.64828733634704
+          "2040": 92.20455788429224
+          "2041": 89.87291235917809
+          "2042": 87.54126683406393
+          "2043": 85.20962130894978
+          "2044": 82.87797578383562
+          "2045": 80.54633025872147
+          "2046": 81.38632144593608
+          "2047": 82.22631263315068
+          "2048": 83.0663038203653
+          "2049": 83.90629500757991
+          "2050": 84.74628619479452
+          "2051": 84.74628619479452
+          "2052": 84.74628619479452
+          "2053": 84.74628619479452
+          "2054": 84.74628619479452
+          "2055": 84.74628619479452
+          "2056": 84.74628619479452
+          "2057": 84.74628619479452
+          "2058": 84.74628619479452
+          "2059": 84.74628619479452
+          "2060": 84.74628619479452
+          "2061": 84.74628619479452
+          "2062": 84.74628619479452
+          "2063": 84.74628619479452
+          "2064": 84.74628619479452
+        # Financial assumptions
+        financial_assumptions:
+          "total income tax rate": 0.2574
+          "capital gains tax rate": 0.15
+          "leverage after tax nominal discount rate": 0.10893
+          "debt equity ratio of initial financing": 0.624788
+          "debt interest rate": 0.050049
